Clean up pipx.
