+++ internal/origins/drivers/pipx/constants.go
@@ -3,4 +3,12 @@ package pipx
 const (
 	defaultVenvRoot = ".local/pipx/venvs"
 	otherVenvRoot   = ".local/share/pipx/venvs"
+
+	fieldName        = "Name"
+	fieldVersion     = "Version"
+	fieldSummary     = "Summary"
+	fieldLicense     = "License"
+	fieldHomepage    = "Home-page"
+	fieldProjectUrl  = "Project-Url"
+	subfieldHomepage = "Homepage"
 )
