+++ internal/origins/drivers/pipx/fetch.go
@@ -1,7 +1,6 @@
 package pipx
 
 import (
-	"bufio"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -9,9 +8,14 @@ import (
 	"qp/internal/origins/shared"
 	"qp/internal/origins/worker"
 	"qp/internal/pkgdata"
-	"strings"
 	"sync"
 )
+
+type PkgMeta struct {
+	Pkg          *pkgdata.PkgInfo
+	DirPath      string
+	SitePkgsPath string
+}
 
 func fetchPackages(venvRoot string, origin string) ([]*pkgdata.PkgInfo, error) {
 	dirs, err := os.ReadDir(venvRoot)
@@ -30,14 +34,7 @@ func fetchPackages(venvRoot string, origin string) ([]*pkgdata.PkgInfo, error) {
 
 		inputChan <- dir
 	}
-
 	close(inputChan)
-
-	type PkgMeta struct {
-		Pkg          *pkgdata.PkgInfo
-		DirPath      string
-		SitePkgsPath string
-	}
 
 	stage1 := worker.RunWorkers(
 		inputChan,
@@ -103,7 +100,7 @@ func fetchPackages(venvRoot string, origin string) ([]*pkgdata.PkgInfo, error) {
 		errChan,
 		&errGroup,
 		func(pMeta *PkgMeta) (*pkgdata.PkgInfo, error) {
-			arch, err := findArchitecture(pMeta.SitePkgsPath)
+			arch, err := inferArchitecture(pMeta.SitePkgsPath)
 			if err != nil {
 				return nil, err
 			}
@@ -121,85 +118,4 @@ func fetchPackages(venvRoot string, origin string) ([]*pkgdata.PkgInfo, error) {
 	}()
 
 	return worker.CollectOutput(stage3, errChan)
-}
-
-func findVersionPath(libRoot string) (string, error) {
-	entries, err := os.ReadDir(libRoot)
-	if err != nil {
-		return "", err
-	}
-
-	for _, e := range entries {
-		if e.IsDir() && e.Name()[:6] == "python" {
-			return filepath.Join(libRoot, e.Name()), nil
-		}
-	}
-	return "", fmt.Errorf("no pythonX.Y found under %s", libRoot)
-}
-
-func findDistPath(sitePkgsPath string, name string) (string, error) {
-	matches, _ := filepath.Glob(filepath.Join(sitePkgsPath, name+"-*.dist-info", "METADATA"))
-	if len(matches) > 0 {
-		return matches[0], nil
-	}
-
-	return "", nil
-}
-
-func findArchitecture(sitePkgsPath string) (string, error) {
-	matches, err := filepath.Glob(filepath.Join(sitePkgsPath, "*.dist-info", "WHEEL"))
-	if err != nil {
-		return "", fmt.Errorf("found no .dist-info directories in %s: %v", sitePkgsPath, err)
-	}
-
-	bestMatch := "any"
-
-	for _, match := range matches {
-		arch, err := parseWheelFile(match)
-		if err != nil {
-			continue
-		}
-
-		parts := strings.Split(arch, "-")
-		suffix := parts[len(parts)-1]
-
-		if suffix != "any" {
-			if !strings.Contains(suffix, "universal") {
-				return suffix, nil
-			}
-
-			bestMatch = suffix
-		}
-	}
-
-	return bestMatch, nil
-}
-
-func parseWheelFile(wheelPath string) (string, error) {
-	file, err := os.Open(wheelPath)
-	if err != nil {
-		return "", err
-	}
-
-	defer file.Close()
-	scanner := bufio.NewScanner(file)
-
-	for scanner.Scan() {
-		line := scanner.Text()
-		if line == "" {
-			break
-		}
-
-		parts := strings.SplitN(line, ":", 2)
-		if len(parts) != 2 {
-			continue
-		}
-
-		key := strings.TrimSpace(parts[0])
-		if key == "Tag" {
-			return strings.TrimSpace(parts[1]), nil
-		}
-	}
-
-	return "", fmt.Errorf("no tag field in wheel file for %s", wheelPath)
 }
