+++ internal/origins/drivers/pipx/parser.go
@@ -2,12 +2,12 @@ package pipx
 
 import (
 	"bufio"
+	"fmt"
 	"os"
 	"qp/internal/pkgdata"
 	"strings"
 )
 
-// TODO: parse line by line
 func parseMetadataFile(metadataPath string) (*pkgdata.PkgInfo, error) {
 	file, err := os.Open(metadataPath)
 	if err != nil {
@@ -36,31 +36,64 @@ func parseMetadataFile(metadataPath string) (*pkgdata.PkgInfo, error) {
 		value := strings.TrimSpace(parts[1])
 
 		switch key {
-		case "Name":
+		case fieldName:
 			pkg.Name = value
-		case "Version":
+		case fieldVersion:
 			pkg.Version = value
-		case "Summary":
+		case fieldSummary:
 			pkg.Description = value
-		case "License":
+		case fieldLicense:
 			pkg.License = value
-		case "Home-page":
+		case fieldHomepage:
 			if homepage == "" {
 				homepage = value
 			}
-		case "Project-URL":
+		case fieldProjectUrl:
 			fieldParts := strings.SplitN(value, ",", 2)
 			if len(fieldParts) != 2 {
 				continue
 			}
 
-			if strings.TrimSpace(fieldParts[0]) == "Homepage" {
+			if strings.TrimSpace(fieldParts[0]) == subfieldHomepage {
 				homepage = strings.TrimSpace(fieldParts[1])
 			}
 		}
 	}
 
+	if err := scanner.Err(); err != nil {
+		return nil, fmt.Errorf("error reading metadata file %s: %w", metadataPath, err)
+	}
+
 	pkg.Url = homepage
 
 	return pkg, nil
+}
+
+func parseWheelFile(wheelPath string) (string, error) {
+	file, err := os.Open(wheelPath)
+	if err != nil {
+		return "", err
+	}
+
+	defer file.Close()
+	scanner := bufio.NewScanner(file)
+
+	for scanner.Scan() {
+		line := scanner.Text()
+		if line == "" {
+			break
+		}
+
+		parts := strings.SplitN(line, ":", 2)
+		if len(parts) != 2 {
+			continue
+		}
+
+		key := strings.TrimSpace(parts[0])
+		if key == "Tag" {
+			return strings.TrimSpace(parts[1]), nil
+		}
+	}
+
+	return "", fmt.Errorf("no tag field in wheel file for %s", wheelPath)
 }
